% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chexpweibull}
\alias{chexpweibull}
\title{Exponentiated Weibull (EW) cumulative hazard function.
http://rpubs.com/FJRubio/EW}
\usage{
chexpweibull(t, lambda, kappa, alpha, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{lambda}{: scale parameter}

\item{kappa}{: shape parameter}

\item{alpha}{: shape parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the EW cumulative hazard function
}
\description{
Exponentiated Weibull (EW) cumulative hazard function.
http://rpubs.com/FJRubio/EW
}
