% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chgamma}
\alias{chgamma}
\title{Gamma (G) Cumulative Hazard Function.}
\usage{
chgamma(t, shape, scale)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}
}
\value{
the value of the G cumulative hazard function
}
\description{
Gamma (G) Cumulative Hazard Function.
}
