% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chggamma}
\alias{chggamma}
\title{Generalised Gamma (GG) Cumulative Hazard Function.
http://rpubs.com/FJRubio/GG}
\usage{
chggamma(t, theta, kappa, delta)
}
\arguments{
\item{t}{: positive argument}

\item{theta}{: scale parameter}

\item{kappa}{: shape parameter}

\item{delta}{: shape parameter}
}
\value{
the value of the GG cumulative hazard function
}
\description{
Generalised Gamma (GG) Cumulative Hazard Function.
http://rpubs.com/FJRubio/GG
}
