% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chllogis}
\alias{chllogis}
\title{Lognormal (LL) Cumulative Hazard Function.}
\usage{
chllogis(t, mu, sigma)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{mu}{: log location parameter}

\item{sigma}{: scale parameter}
}
\value{
the value of the LL cumulative hazard function
}
\description{
Lognormal (LL) Cumulative Hazard Function.
}
