% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chlnorm}
\alias{chlnorm}
\title{Lognormal (LN) Cumulative Hazard Function.}
\usage{
chlnorm(t, mu, sigma)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{mu}{: log location parameter}

\item{sigma}{: scale parameter}
}
\value{
the value of the LN cumulative hazard function
}
\description{
Lognormal (LN) Cumulative Hazard Function.
}
