% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{chpgw}
\alias{chpgw}
\title{Power Generalised Weibull (PGW) cumulative hazard function.
http://rpubs.com/FJRubio/PGW}
\usage{
chpgw(t, eta, nu, delta)
}
\arguments{
\item{t}{: positive argument}

\item{eta}{: scale parameter}

\item{nu}{: shape parameter}

\item{delta}{: shape parameter}
}
\value{
the value of the PGW cumulative hazard function
}
\description{
Power Generalised Weibull (PGW) cumulative hazard function.
http://rpubs.com/FJRubio/PGW
}
