% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{dggamma}
\alias{dggamma}
\title{Generalised Gamma (GG) Probability Density Function.
http://rpubs.com/FJRubio/GG}
\usage{
dggamma(t, theta, kappa, delta, log = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{theta}{: scale parameter}

\item{kappa}{: shape parameter}

\item{delta}{: shape parameter}
}
\value{
the value of the GG probability density function
}
\description{
Generalised Gamma (GG) Probability Density Function.
http://rpubs.com/FJRubio/GG
}
