% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{hexpweibull}
\alias{hexpweibull}
\title{Exponentiated Weibull (EW) hazard function.
http://rpubs.com/FJRubio/EW}
\usage{
hexpweibull(t, lambda, kappa, alpha, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{lambda}{: scale parameter}

\item{kappa}{: shape parameter}

\item{alpha}{: shape parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the EW hazard function
}
\description{
Exponentiated Weibull (EW) hazard function.
http://rpubs.com/FJRubio/EW
}
