% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{hgamma}
\alias{hgamma}
\title{Gamma (G) Hazard Function.}
\usage{
hgamma(t, shape, scale, log = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}
}
\value{
the value of the G hazard function
}
\description{
Gamma (G) Hazard Function.
}
