% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{hllogis}
\alias{hllogis}
\title{Log-logistic (LL) Hazard Function.}
\usage{
hllogis(t, mu, sigma, log = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{mu}{: log location parameter}

\item{sigma}{: scale parameter}
}
\value{
the value of the LL hazard function
}
\description{
Log-logistic (LL) Hazard Function.
}
