% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{hlnorm}
\alias{hlnorm}
\title{Lognormal (LN) Hazard Function.}
\usage{
hlnorm(t, mu, sigma, log = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{mu}{: log location parameter}

\item{sigma}{: scale parameter}
}
\value{
the value of the LN hazard function
}
\description{
Lognormal (LN) Hazard Function.
}
