% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{hpgw}
\alias{hpgw}
\title{Power Generalised Weibull (PGW) hazard function.
http://rpubs.com/FJRubio/PGW}
\usage{
hpgw(t, eta, nu, delta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{eta}{: scale parameter}

\item{nu}{: shape parameter}

\item{delta}{: shape parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the PGW hazard function
}
\description{
Power Generalised Weibull (PGW) hazard function.
http://rpubs.com/FJRubio/PGW
}
