% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{pggamma}
\alias{pggamma}
\title{Generalised Gamma (GG) Cumulative Distribution Function.
http://rpubs.com/FJRubio/GG}
\usage{
pggamma(t, theta, kappa, delta, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log.p: log scale (TRUE or FALSE)}

\item{theta}{: scale parameter}

\item{kappa}{: shape parameter}

\item{delta}{: shape parameter}
}
\value{
the value of the GG cumulative distribution function
}
\description{
Generalised Gamma (GG) Cumulative Distribution Function.
http://rpubs.com/FJRubio/GG
}
