% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHSurv.R
\name{sggamma}
\alias{sggamma}
\title{Generalised Gamma (GG) Survival Function.
http://rpubs.com/FJRubio/GG}
\usage{
sggamma(t, theta, kappa, delta, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument
#' @param log: log scale (TRUE or FALSE)}

\item{theta}{: scale parameter}

\item{kappa}{: shape parameter}

\item{delta}{: shape parameter}
}
\value{
the value of the GG survival function
}
\description{
Generalised Gamma (GG) Survival Function.
http://rpubs.com/FJRubio/GG
}
